# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vulnerability Assistance Platform API",
    "version": "1.0.0",
    "description": "API para plataforma de acompanhamento de pessoas em vulnerabilidade urbana",
    "contact": {
      "name": "API Support",
      "email": "support@vulnerability-platform.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT válido por 5 dias"
      }
    },
    "schemas": {
...
      "Person": {
        "type": "object",
        "required": [
          "nome"
        ],
        "properties": {
          "person_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID único da pessoa (auto-gerado)",
            "readOnly": true
          },
          "id_interno": {
            "type": "string",
            "maxLength": 50,
            "readOnly": true,
            "description": "ID interno único gerado automaticamente"
          },
          "nome": {
            "type": "string",
            "maxLength": 255,
            "description": "Nome completo"
          },
          "data_nascimento": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Data de nascimento"
          },
          "genero": {
            "type": "string",
            "maxLength": 50,
            "nullable": true,
            "description": "Gênero"
          },
          "endereco": {
            "type": "string",
            "nullable": true,
            "description": "Endereço completo"
          },
          "contato": {
            "type": "object",
            "nullable": true,
            "description": "Informações de contato (JSON)"
          },
          "dados_sensivel": {
            "type": "object",
            "nullable": true,
            "description": "Dados sensíveis (JSON)"
          },
          "data_cadastro": {
            "type": "string",
            "format": "date-time",
            "description": "Data de cadastro"
          },
          "origem": {
            "type": "string",
            "maxLength": 255,
            "nullable": true,
            "description": "Origem da pessoa"
          }
        }
      },
      "DynamicMetadata": {
        "type": "object",
        "required": [
          "person_id",
          "tipo_metadado",
          "categoria",
          "descricao"
        ],
        "properties": {
          "metadata_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID único do metadado (auto-gerado)",
            "readOnly": true
          },
          "person_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID da pessoa"
          },
          "tipo_metadado": {
            "type": "string",
            "enum": [
              "experiencia",
              "acontecimento"
            ],
            "description": "Tipo do metadado"
          },
          "categoria": {
            "type": "string",
            "maxLength": 255,
            "description": "Categoria do metadado"
          },
          "sub_categoria": {
            "type": "string",
            "maxLength": 255,
            "nullable": true,
            "description": "Subcategoria"
          },
          "descricao": {
            "type": "string",
            "description": "Descrição detalhada"
          },
          "data_ocorrencia": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Data da ocorrência"
          },
          "detalhes_json": {
            "type": "object",
            "nullable": true,
            "description": "Detalhes adicionais (JSON)"
          },
          "data_registro": {
            "type": "string",
            "format": "date-time",
            "description": "Data de registro"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica se a operação foi bem-sucedida"
          },
          "data": {
            "type": "object",
            "description": "Dados retornados"
          },
          "message": {
            "type": "string",
            "description": "Mensagem de resposta"
          },
          "count": {
            "type": "integer",
            "description": "Número de itens (para listas)"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Mensagem de erro"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/metadata": {
      "post": {
        "summary": "Registrar novo metadado",
        "tags": [
          "Metadata"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DynamicMetadata"
              },
              "example": {
                "person_id": "uuid-da-pessoa",
                "tipo_metadado": "experiencia",
                "categoria": "Educação",
                "sub_categoria": "Alfabetização",
                "descricao": "Participou do programa de alfabetização",
                "detalhes_json": {
                  "nivel": "básico",
                  "duracao": "6 meses"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Metadado registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar metadados",
        "tags": [
          "Metadata"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "person_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtrar por ID da pessoa"
          },
          {
            "in": "query",
            "name": "tipo_metadado",
            "schema": {
              "type": "string",
              "enum": [
                "experiencia",
                "acontecimento"
              ]
            },
            "description": "Filtrar por tipo de metadado"
          },
          {
            "in": "query",
            "name": "categoria",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por categoria"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de metadados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/metadata/{id}": {
      "get": {
        "summary": "Buscar metadado por ID",
        "tags": [
          "Metadata"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do metadado"
          }
        ],
        "responses": {
          "200": {
            "description": "Metadado encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Metadado não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar metadado",
        "tags": [
          "Metadata"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do metadado"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DynamicMetadata"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Metadado atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remover metadado",
        "tags": [
          "Metadata"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do metadado"
          }
        ],
        "responses": {
          "200": {
            "description": "Metadado removido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Metadado não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/persons": {
      "post": {
        "summary": "Cadastrar nova pessoa",
        "tags": [
          "Persons"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              },
              "example": {
                "id_interno": "PES001",
                "nome": "Maria Silva",
                "genero": "Feminino",
                "endereco": "Rua das Flores, 123",
                "contato": {
                  "telefone": "11999999999",
                  "email": "maria@example.com"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pessoa cadastrada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar pessoas",
        "tags": [
          "Persons"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nome",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por nome"
          },
          
          {
            "in": "query",
            "name": "genero",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por gênero"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pessoas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/persons/{id}": {
      "get": {
        "summary": "Buscar pessoa por ID",
        "tags": [
          "Persons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID da pessoa"
          }
        ],
        "responses": {
          "200": {
            "description": "Pessoa encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pessoa não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar dados da pessoa",
        "tags": [
          "Persons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID da pessoa"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pessoa atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remover pessoa do sistema",
        "tags": [
          "Persons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID da pessoa"
          }
        ],
        "responses": {
          "200": {
            "description": "Pessoa removida com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pessoa não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}